udpating library package for RxJava2.x
