+++ library/src/main/java/com/github/pwittchen/reactivebeacons/library/rx2/Filter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.pwittchen.reactivebeacons.library;
+package com.github.pwittchen.reactivebeacons.library.rx2;
 
 
 import io.reactivex.functions.Function;
